requirements.txt
letta-client
python-dotenv
rich